Refactoring: move models to dedicated folder
