+++ bambooai/__init__.py
@@ -7,16 +7,6 @@ from . import qa_retrieval
 from . import models
 from . import google_search
 from . import reg_ex
-from . import local_models
-from . import groq_models
-from . import openai_models
-from . import deepseek_models
-from . import ollama_models
-from . import vllm_models
-from . import gemini_models
-from . import anthropic_models
-from . import mistral_models
-from . import openrouter_models
 from . import log_manager
 from . import storage_manager
 from . import output_manager
