+++ requirements.txt
@@ -28,3 +28,4 @@ scipy
 scikit-learn
 geopandas
 yfinance
+opentelemetry-distro
